generator/requirements.txt
-i https://pypi.org/simple
aiofiles==23.2.1; python_version >= '3.7'
aiohttp==3.8.6; python_version >= '3.6'
aiosignal==1.3.1; python_version >= '3.7'
altair==5.1.2; python_version >= '3.8'
antlr4-python3-runtime==4.9.3
anyio==3.7.1; python_version >= '3.7'
asttokens==2.4.0
async-timeout==4.0.3; python_version >= '3.7'
attrs==23.1.0; python_version >= '3.7'
audiocraft==1.0.0; python_full_version >= '3.8.0'
audioread==3.0.1; python_version >= '3.6'
av==10.0.0
backcall==0.2.0
blis==0.7.11
catalogue==2.0.10; python_version >= '3.6'
certifi==2023.7.22; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
charset-normalizer==3.3.1; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
cloudpickle==3.0.0; python_version >= '3.8'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
colorlog==6.7.0; python_version >= '3.6'
confection==0.1.3; python_version >= '3.6'
contourpy==1.1.1; python_version >= '3.8'
cycler==0.12.1; python_version >= '3.8'
cymem==2.0.8
decorator==5.1.1; python_version >= '3.5'
demucs==4.0.1; python_full_version >= '3.8.0'
diskcache==5.6.3; python_version >= '3'
docopt==0.6.2
dora-search==0.1.12; python_full_version >= '3.7.0'
einops==0.7.0; python_version >= '3.8'
emoji==2.8.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
encodec==0.1.1; python_full_version >= '3.8.0'
executing==2.0.0
fastapi==0.104.0; python_version >= '3.8'
ffmpy==0.3.1
filelock==3.12.4; python_version >= '3.8'
flashy==0.0.2; python_full_version >= '3.8.0'
fonttools==4.43.1; python_version >= '3.8'
frozenlist==1.4.0; python_version >= '3.8'
fsspec==2023.10.0; python_version >= '3.8'
gradio==3.50.2; python_version >= '3.8'
gradio-client==0.6.1; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
halo==0.0.31; python_version >= '3.4'
httpcore==0.18.0; python_version >= '3.8'
httpx==0.25.0; python_version >= '3.8'
huggingface-hub==0.17.3; python_full_version >= '3.8.0'
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.4; python_version >= '3.5'
importlib-resources==6.1.0; python_version >= '3.8'
ipython==8.16.1; python_version >= '3.9'
jedi==0.19.1; python_version >= '3.6'
jinja2==3.1.2; python_version >= '3.7'
joblib==1.3.2; python_version >= '3.7'
jsonschema==4.19.1; python_version >= '3.8'
jsonschema-specifications==2023.7.1; python_version >= '3.8'
julius==0.2.7; python_full_version >= '3.6.0'
kiwisolver==1.4.5; python_version >= '3.7'
lameenc==1.6.1
langcodes==3.3.0; python_version >= '3.6'
lazy-loader==0.3; python_version >= '3.7'
librosa==0.10.1; python_version >= '3.7'
lightning-utilities==0.9.0; python_version >= '3.7'
llama-cpp-python==0.2.11; python_version >= '3.8'
llvmlite==0.41.1; python_version >= '3.8'
log-symbols==0.0.14
markupsafe==2.1.3; python_version >= '3.7'
matplotlib==3.8.0; python_version >= '3.9'
matplotlib-inline==0.1.6; python_version >= '3.5'
mpmath==1.3.0
msgpack==1.0.7; python_version >= '3.8'
multidict==6.0.4; python_version >= '3.7'
murmurhash==1.0.10; python_version >= '3.6'
networkx==3.2; python_version >= '3.9'
num2words==0.5.13
numba==0.58.1; python_version >= '3.8'
numpy==1.26.1; python_version < '3.13' and python_version >= '3.9'
nvidia-cublas-cu12==12.1.3.1; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cuda-cupti-cu12==12.1.105; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cuda-nvrtc-cu12==12.1.105; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cuda-runtime-cu12==12.1.105; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cudnn-cu12==8.9.2.26; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cufft-cu12==11.0.2.54; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-curand-cu12==10.3.2.106; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cusolver-cu12==11.4.5.107; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cusparse-cu12==12.1.0.106; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-nccl-cu12==2.18.1; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-nvjitlink-cu12==12.3.52; python_version >= '3'
nvidia-nvtx-cu12==12.1.105; platform_system == 'Linux' and platform_machine == 'x86_64'
omegaconf==2.3.0; python_version >= '3.6'
openai==0.28.1; python_full_version >= '3.7.1'
openunmix==1.2.1; python_version >= '3.6'
orjson==3.9.9; python_version >= '3.8'
packaging==23.2; python_version >= '3.7'
pandas==2.1.1; python_version >= '3.9'
parso==0.8.3; python_version >= '3.6'
pathy==0.10.3; python_version >= '3.6'
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
pillow==10.1.0; python_version >= '3.8'
platformdirs==3.11.0; python_version >= '3.7'
pooch==1.7.0; python_version >= '3.7'
preshed==3.0.9; python_version >= '3.6'
prompt-toolkit==3.0.39; python_full_version >= '3.7.0'
protobuf==4.24.4; python_version >= '3.7'
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.13; python_version >= '3.7'
pydub==0.25.1
pygments==2.16.1; python_version >= '3.7'
pyparsing==3.1.1; python_full_version >= '3.6.8'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0; python_version >= '3.8'
python-multipart==0.0.6; python_version >= '3.7'
pytz==2023.3.post1
pyyaml==6.0.1; python_version >= '3.6'
referencing==0.30.2; python_version >= '3.8'
regex==2023.10.3; python_version >= '3.7'
requests==2.31.0; python_version >= '3.7'
retrying==1.3.4
rpds-py==0.10.6; python_version >= '3.8'
ruamel.yaml==0.17.40; python_version >= '3'
ruamel.yaml.clib==0.2.8; python_version < '3.13' and platform_python_implementation == 'CPython'
safetensors==0.4.0; python_version >= '3.7'
scikit-learn==1.3.1; python_version >= '3.8'
scipy==1.11.3; python_version < '3.13' and python_version >= '3.9'
semantic-version==2.10.0; python_version >= '2.7'
sentencepiece==0.1.99
setuptools==68.2.2; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smart-open==6.4.0; python_version >= '3.6' and python_version < '4.0'
sniffio==1.3.0; python_version >= '3.7'
soundfile==0.12.1
soxr==0.3.7; python_version >= '3.6'
spacy==3.5.2; python_version >= '3.6'
spacy-legacy==3.0.12; python_version >= '3.6'
spacy-loggers==1.0.5; python_version >= '3.6'
spinners==0.0.24
srsly==2.4.8; python_version >= '3.6'
stack-data==0.6.3
starlette==0.27.0; python_version >= '3.7'
submitit==1.5.0; python_version >= '3.8'
sympy==1.12; python_version >= '3.8'
termcolor==2.3.0; python_version >= '3.7'
thinc==8.1.12; python_version >= '3.6'
threadpoolctl==3.2.0; python_version >= '3.8'
tokenizers==0.14.1; python_version >= '3.7'
toolz==0.12.0; python_version >= '3.5'
torch==2.1.0; python_full_version >= '3.8.0'
torchaudio==2.1.0
torchmetrics==1.2.0; python_version >= '3.8'
tqdm==4.66.1; python_version >= '3.7'
traitlets==5.11.2; python_version >= '3.8'
transformers==4.34.1; python_full_version >= '3.8.0'
treetable==0.2.5; python_full_version >= '3.6.0'
triton==2.1.0; platform_system == 'Linux' and platform_machine == 'x86_64'
typer==0.7.0; python_version >= '3.6'
typing-extensions==4.8.0; python_version >= '3.8'
tzdata==2023.3; python_version >= '2'
urllib3==2.0.7; python_version >= '3.7'
uvicorn==0.23.2; python_version >= '3.8'
wasabi==1.1.2; python_version >= '3.6'
wcwidth==0.2.8
websockets==11.0.3; python_version >= '3.7'
xformers==0.0.22.post4; python_version >= '3.7'
yarl==1.9.2; python_version >= '3.7'
